requirements.txt
batchbald_redux==2.0.5
hydra-core==1.3.2
numpy==1.25.1
omegaconf==2.3.0
pandas==2.2.3
peft==0.13.2
scikit_learn==1.5.2
scipy==1.14.1
tqdm==4.65.0
trl==0.11.4
ucimlrepo==0.0.7
unsloth==2024.10.7
wandb==0.18.5
